refactor: simplify articles previews and remove tags selector
